Changing FeedbackRadius type from int to FLOAT. Changed some GalaxyParticle files too. An effort to rectify star issues.

+++ src/enzo/ActiveParticle_AccretingParticle.C
@@ -304,7 +304,7 @@ int ActiveParticleType_AccretingParticle::BeforeEvolveLevel
 }
 
 
-grid* ConstructFeedbackZone(ActiveParticleType* ThisParticle, int FeedbackRadius,
+grid* ConstructFeedbackZone(ActiveParticleType* ThisParticle, FLOAT FeedbackRadius,
 			     FLOAT dx, HierarchyEntry** Grids, int NumberOfGrids,
 			     int SendField);
